lab3/requirements.txt
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
opencv-python==4.7.0.72
pandas==2.0.0
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
Werkzeug==2.2.3
